1.CREATE Protect APP

+++ USER/Sys/ringbuffer.h
@@ -3,21 +3,22 @@
 #define __RINGBUFFER_H__
 
 #include "stdbool.h"
+#include <stdio.h>
 
 struct ring_buffer {
-	unsigned int in; /*写指针*/
-	unsigned int out; /*读指针*/
-	unsigned int size;  /*环形缓冲区大小*/
-	unsigned char *buf; /*环形缓冲区*/
+	unsigned short in;
+	unsigned short out;
+	unsigned short size;
+	unsigned char *buf;
 };
 
-bool RingBufferInit(struct ring_buffer *ring_buffer_p, unsigned char *pool, unsigned int buf_size);
+bool RingBufferInit(struct ring_buffer *ring_buffer_p, unsigned char *pool, unsigned short buf_size);
 bool RingBufferEmpty(struct ring_buffer *ring_buffer_p);
 bool RingBufferFull(struct ring_buffer *ring_buffer_p);
-unsigned int GetRingBufferLen(struct ring_buffer *ring_buffer_p);
+unsigned short GetRingBufferLen(struct ring_buffer *ring_buffer_p);
 void RingBufferClear(struct ring_buffer *ring_buffer_p);
 void RingBufferDel(struct ring_buffer *ring_buffer_p);
-int RingBufferPut(struct ring_buffer *ring_buffer_p, unsigned char *buf, unsigned int len);
-unsigned int RingBufferGet(struct ring_buffer *ring_buffer_p, unsigned char *buf, unsigned int buf_len);
+int RingBufferPut(struct ring_buffer *ring_buffer_p, unsigned char *buf, unsigned short len);
+unsigned short RingBufferGet(struct ring_buffer *ring_buffer_p, unsigned char *buf, unsigned short buf_len);
 
 #endif
